refactor(react-antd): history降4.10.1版本，拆分config/router解决路由问题
