+++ packages/react-antd-boilerplate/package.json
@@ -12,8 +12,8 @@
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "connected-react-router": "^6.8.0",
-    "history": "^5.0.0",
     "js-md5": "^0.7.3",
+    "history": "^4.10.1",
     "moment": "^2.27.0",
     "node-sass": "^4.14.1",
     "prop-types": "^15.7.2",
