+++ packages/react-antd-boilerplate/src/App.js
@@ -1,24 +1,25 @@
 /*
  * @Autor: yugeStrive
  * @Date: 2020-07-07 17:02:40
- * @LastEditTime: 2020-07-14 18:39:33
+ * @LastEditTime: 2020-07-16 15:21:47
  * @Description: 主页面
  */
 
 import React, { Component } from 'react'
 import { Provider } from 'react-redux'
-import RouterView, { history } from './router'
-import { BrowserRouter as Router } from 'react-router-dom'
-import routeLists from './views/root.routes'
+import { ConnectedRouter } from 'connected-react-router'
+import RouterView from '@/router'
+import { history } from '@/router/history'
+import routeList from './views/root.routes'
 import store from '@/store'
 
 class App extends Component {
   render() {
     return (
       <Provider store={store}>
-        <Router history={history}>
-          <RouterView routeLists={routeLists}></RouterView>
-        </Router>
+        <ConnectedRouter history={history}>
+          <RouterView routeList={routeList}></RouterView>
+        </ConnectedRouter>
       </Provider>
     )
   }
