+++ packages/react-antd-boilerplate/src/config/envs.js
@@ -1,0 +1,13 @@
+/*
+ * @Autor: yugeStrive
+ * @Date: 2020-07-16 14:39:55
+ * @LastEditTime: 2020-07-16 14:41:46
+ * @Description: 路由模式
+ */
+
+export default {
+  // 路由模式
+  ROUTE_MODE: 'hash',
+  // vconsole 插件开关，设置为 true 也仅在开发环境下生效（PC端项目请忽略）
+  VCONSOLE_ENABLE: true,
+}
