+++ packages/react-antd-boilerplate/src/router/index.js
@@ -1,25 +1,20 @@
 /*
  * @Author: yugeStrive
  * @Date: 2020-07-12 10:22:30
- * @LastEditTime: 2020-07-14 18:41:35
+ * @LastEditTime: 2020-07-16 15:21:29
  * @Description: 全局路由跳转方式
  */
 
 import React, { Component } from 'react'
 import { Route, Switch, Redirect } from 'react-router-dom'
-import { createBrowserHistory } from 'history'
 import config from '@/config'
-// import { connect } from 'react-redux'
-// import { routerActions } from 'connected-react-router'
-
-export const history = createBrowserHistory()
 
 class RouterView extends Component {
   render() {
     return (
       <div>
           <Switch>
-            {this.props.routeLists.map((item) => {
+            {this.props.routeList.map((item) => {
               return (
                 <Route
                   exact={item.exact}
@@ -29,7 +24,7 @@ class RouterView extends Component {
                 />
               )
             })}
-            <Redirect from="/*" to={config.UNAUTHORIZED_REDIRECT_PATH} />
+            <Redirect from="/*" to={config.SERVICES.UNAUTHORIZED_REDIRECT_PATH} />
           </Switch>
       </div>
     )
