+++ packages/react-antd-boilerplate/src/views/root.routes.js
@@ -1,14 +1,14 @@
 /*
  * @Author: yugeStrive
  * @Date: 2020-07-12 10:22:30
- * @LastEditTime: 2020-07-13 14:16:50
+ * @LastEditTime: 2020-07-16 15:22:01
  * @Description: views路由数据项
  */
 
 import ReactPage from './reactPage/index'
 import Login from './login/index'
 
-const routeLists = [
+const routeList = [
     {
         path: '/',
         component: ReactPage,
@@ -21,4 +21,4 @@ const routeLists = [
     }
 ]
 
-export default routeLists
+export default routeList
